# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_adapters_http_equipment.CalibrationEventResponse:
    properties:
      acceptance_value:
        type: string
      calibrate_value:
        type: string
      calibrated_at:
        type: string
      calibration_type:
        type: string
      equipment_id:
        type: string
      id:
        type: integer
      next_calibration_due:
        type: string
      notes:
        type: string
      performed_by:
        type: string
      result:
        type: string
    type: object
  internal_adapters_http_equipment.CalibrationScheduleResponse:
    properties:
      equipment_id:
        type: string
      id:
        type: integer
      interval_months:
        type: integer
      label:
        type: string
      next_due_date:
        type: string
    type: object
  internal_adapters_http_equipment.CreateCalibrationScheduleRequest:
    properties:
      interval_months:
        type: integer
      label:
        type: string
      next_due_date:
        type: string
    required:
    - label
    - next_due_date
    type: object
  internal_adapters_http_equipment.CreateEquipmentRequest:
    properties:
      category:
        type: string
      installation_date:
        type: string
      location_id:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      next_calibration_due:
        type: string
      serial_number:
        type: string
      type_code:
        type: string
      vendor_id:
        type: string
    required:
    - name
    - next_calibration_due
    - type_code
    type: object
  internal_adapters_http_equipment.EquipmentResponse:
    properties:
      calibration_pct:
        type: integer
      category:
        type: string
      id:
        type: string
      installation_date:
        type: string
      last_calibrated_at:
        type: string
      location_id:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      next_calibration_due:
        type: string
      serial_number:
        type: string
      status:
        type: string
      type_code:
        type: string
      usage_hours:
        type: integer
      vendor_id:
        type: string
    type: object
  internal_adapters_http_equipment.RecordCalibrationRequest:
    properties:
      acceptance_value:
        type: string
      calibrate_value:
        type: string
      calibration_type:
        type: string
      next_calibration_due:
        type: string
      notes:
        type: string
      result:
        enum:
        - pass
        - fail
        type: string
    required:
    - next_calibration_due
    type: object
  internal_adapters_http_equipment.UpdateCalibrationScheduleRequest:
    properties:
      clear_interval:
        type: boolean
      interval_months:
        type: integer
      label:
        type: string
      next_due_date:
        type: string
    type: object
  internal_adapters_http_equipment.UpdateEquipmentRequest:
    properties:
      category:
        type: string
      clear_installation_date:
        type: boolean
      installation_date:
        type: string
      location_id:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      serial_number:
        type: string
      type_code:
        type: string
      vendor_id:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API สำหรับ Thanes LIMS backend (Auth, Sample/Chain-of-Custody,
    Test Result, Equipment, Inventory/Purchase Order, Document, Environment, Notification,
    Audit)
  title: Thanes LIMS Backend API
  version: "1.0"
paths:
...
  /calibration-results:
    get:
      description: รายการผลสอบเทียบข้ามทุกอุปกรณ์ พร้อม search bar (requirement 2.2.1)
      parameters:
      - description: ค้นหา (รหัส/ชื่ออุปกรณ์, ผู้สอบเทียบ, ประเภท, หมายเหตุ)
        in: query
        name: q
        type: string
      - description: กรองตามอุปกรณ์
        in: query
        name: equipment_id
        type: string
      - description: pass หรือ fail
        in: query
        name: result
        type: string
      - description: วันสอบเทียบตั้งแต่ (RFC3339)
        in: query
        name: from
        type: string
      - description: วันสอบเทียบถึง (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_equipment.CalibrationEventResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ค้นหาผลการสอบเทียบทั้งหมด
      tags:
      - equipment
...
  /equipment:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_equipment.EquipmentResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการอุปกรณ์ทั้งหมด
      tags:
      - equipment
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลอุปกรณ์
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_equipment.CreateEquipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.EquipmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: เพิ่มอุปกรณ์ใหม่
      tags:
      - equipment
  /equipment/{id}:
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.EquipmentResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลอุปกรณ์ตาม ID
      tags:
      - equipment
    patch:
      consumes:
      - application/json
      description: แก้ไขแบบบางส่วน — ส่งเฉพาะ field ที่ต้องการเปลี่ยน (วันสอบเทียบแก้ผ่าน
        /calibration เท่านั้น)
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่แก้ไข
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_equipment.UpdateEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.EquipmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลอุปกรณ์
      tags:
      - equipment
  /equipment/{id}/calibration:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      - description: วันสอบเทียบครั้งถัดไป
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_equipment.RecordCalibrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.EquipmentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: บันทึกการสอบเทียบอุปกรณ์
      tags:
      - equipment
  /equipment/{id}/calibration-events:
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_equipment.CalibrationEventResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ประวัติการสอบเทียบของอุปกรณ์
      tags:
      - equipment
  /equipment/{id}/calibration-schedules:
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_equipment.CalibrationScheduleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการตารางสอบเทียบของอุปกรณ์
      tags:
      - equipment
    post:
      consumes:
      - application/json
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลตารางสอบเทียบ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_equipment.CreateCalibrationScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.CalibrationScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: เพิ่มตารางสอบเทียบให้อุปกรณ์
      tags:
      - equipment
  /equipment/{id}/calibration-schedules/{scheduleId}:
    delete:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Calibration Schedule ID
        in: path
        name: scheduleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ลบตารางสอบเทียบ
      tags:
      - equipment
    patch:
      consumes:
      - application/json
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Calibration Schedule ID
        in: path
        name: scheduleId
        required: true
        type: integer
      - description: ข้อมูลที่แก้ไข
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_equipment.UpdateCalibrationScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_equipment.CalibrationScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: แก้ไขตารางสอบเทียบ
      tags:
      - equipment
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคและ JWT access token เช่น "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"